fix: Fix filter test endpoint type error
- Changed execution_time_ms from int to float in FilterTestResponse
- Fixes TypeError when filter executor returns float milliseconds
- Resolves CORS error that was caused by endpoint crashing

+++ backend/app/api/v1/endpoints/widget_filters.py
@@ -54,7 +54,7 @@ class FilterTestResponse(BaseModel):
     success: bool
     row_count: int
     original_count: int
-    execution_time_ms: int
+    execution_time_ms: float
     sample_data: Optional[List[Dict]] = None
     error: Optional[str] = None
